fix sync 2
